+++ podcasts/Onboarding/OnboardingHostingViewController.swift
@@ -31,10 +31,15 @@ class OnboardingHostingViewController<Content>: UIHostingController<Content>, UI
         }
     }
 
+    override var supportedInterfaceOrientations: UIInterfaceOrientationMask {
+        .portrait
+    }
+
     override func viewDidLoad() {
         super.viewDidLoad()
         presentationController?.delegate = self
         navigationController?.presentationController?.delegate = self
+
         updateNavigationBarStyle(animated: false)
 
         navigationItem.backButtonDisplayMode = .minimal