+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+7.30
+-----
+
+
 7.29
 -----
+- Fixed Edit Folder screen layout (#53)
 - Fixed an issue where episodes wouldn't resume downloading on next launch if the app was force quit (#472)
 - Fixed an issue where the podcast artwork wasn't appearin in the Now Playing Siri suggestions (#579)
 - Display gravatar under Profile (#584)
@@ -9,6 +14,7 @@
 - Updates the onboarding and login process: (#548)
 - Updates the end of year calculation to include the month of December (#554)
 - Updates the End of Year stats designs
+- Fixes an issue where the login button is not accesible on very small devices (#597)
 
 7.27
 -----